I completed this Unity project as part of an interview.
The task was to create a new shader, then instantiate 1000 random game object prefabs with it, then create a screenspace UI with a button to further instantiate, then make a Lambert Shader with GPU instancing to speed it up.  Not sure if I did the last part right... any comments are appreciated!